Extract getDegreeRewriter out from one file again
