+++ community/cypher/cypher-compiler-3.0/src/main/scala/org/neo4j/cypher/internal/compiler/v3_0/ast/rewriters/projectNamedPaths.scala
@@ -19,11 +19,10 @@
  */
 package org.neo4j.cypher.internal.compiler.v3_0.ast.rewriters
 
-import org.neo4j.cypher.internal.frontend.v3_0.Foldable._
-import org.neo4j.cypher.internal.compiler.v3_0._
-import org.neo4j.cypher.internal.frontend.v3_0.ast._
 import org.neo4j.cypher.internal.compiler.v3_0.planner.CantHandleQueryException
-import org.neo4j.cypher.internal.frontend.v3_0.{Ref, topDown, Rewriter}
+import org.neo4j.cypher.internal.frontend.v3_0.Foldable._
+import org.neo4j.cypher.internal.frontend.v3_0.ast._
+import org.neo4j.cypher.internal.frontend.v3_0.{Ref, Rewriter, topDown}
 
 import scala.annotation.tailrec
 
@@ -109,7 +108,6 @@ case object projectNamedPaths extends Rewriter {
 
     case projection: With =>
       (acc, children) =>
-        val projectedVariables = projection.returnItems.items.flatMap(_.alias).toSet
         val projectedAcc = projection.returnItems.items.map(_.expression).foldLeft(acc) {
           (acc, expr) => acc.withVariableRewritesForExpression(expr)
         }
